api/requirements.txt
typing-extensions

gql
# Dependencies for the gql requests transport.
requests-toolbelt<1,>=0.9.1
urllib3>=1.26
requests<3,>=2.26

# Lagoon token validation.
pyjwt
